refactor(WavFileLoader): check subformat type
